Add shellcmd.cc, modify arp_scan cmd.
