+++ source/kernel/main.cc
@@ -91,613 +91,12 @@ static uint32_t rnd_next = 1;
 
 #include <dev/storage/ahci/ahci-raph.h>
 AhciChannel *g_channel = nullptr;
-
-void register_membench2_callout();
-
-static void halt(int argc, const char *argv[]) { acpi_ctrl->Shutdown(); }
-
-static void reset(int argc, const char *argv[]) { acpi_ctrl->Reset(); }
-
-static void lspci(int argc, const char *argv[]) {
-  MCFG *mcfg = acpi_ctrl->GetMCFG();
-  if (mcfg == nullptr) {
-    gtty->Printf("[Pci] error: could not find MCFG table.\n");
-    return;
-  }
-
-  const char *search = nullptr;
-  if (argc > 2) {
-    gtty->Printf("[lspci] error: invalid argument\n");
-    return;
-  } else if (argc == 2) {
-    search = argv[1];
-  }
-
-  PciData::Table table;
-  table.Init();
-
-  for (int i = 0;
-       i * sizeof(MCFGSt) < mcfg->header.Length - sizeof(ACPISDTHeader); i++) {
-    if (i == 1) {
-      gtty->Printf("[Pci] info: multiple MCFG tables.\n");
-      break;
-    }
-    for (int j = mcfg->list[i].pci_bus_start; j <= mcfg->list[i].pci_bus_end;
-         j++) {
-      for (int k = 0; k < 32; k++) {
-        int maxf =
-            ((pci_ctrl->ReadPciReg<uint16_t>(j, k, 0, PciCtrl::kHeaderTypeReg) &
-              PciCtrl::kHeaderTypeRegFlagMultiFunction) != 0)
-                ? 7
-                : 0;
-        for (int l = 0; l <= maxf; l++) {
-          if (pci_ctrl->ReadPciReg<uint16_t>(j, k, l, PciCtrl::kVendorIDReg) ==
-              0xffff) {
-            continue;
-          }
-          table.Search(j, k, l, search);
-        }
-      }
-    }
-  }
-}
-
-static bool parse_ipaddr(const char *c, uint8_t *addr) {
-  int i = 0;
-  while (true) {
-    if (i != 3 && *c == '.') {
-      i++;
-    } else if (i == 3 && *c == '\0') {
-      return true;
-    } else if (*c >= '0' && *c <= '9') {
-      addr[i] *= 10;
-      addr[i] += *c - '0';
-    } else {
-      return false;
-    }
-    c++;
-  }
-}
-
-static void setip(int argc, const char *argv[]) {
-  if (argc < 3) {
-    gtty->Printf("invalid argument\n");
-    return;
-  }
-  NetDevCtrl::NetDevInfo *info = netdev_ctrl->GetDeviceInfo(argv[1]);
-  if (info == nullptr) {
-    gtty->Printf("no ethernet interface(%s).\n", argv[1]);
-    return;
-  }
-  NetDev *dev = info->device;
-
-  uint8_t addr[4] = {0, 0, 0, 0};
-  if (!strcmp(argv[2], "static")) {
-    if (argc != 4) {
-      gtty->Printf("invalid argument\n");
-      return;
-    }
-    if (!parse_ipaddr(argv[3], addr)) {
-      gtty->Printf("invalid ip v4 addr.\n");
-      return;
-    }
-    dev->AssignIpv4Address(*(reinterpret_cast<uint32_t *>(addr)));
-  } else if (!strcmp(argv[2], "dhcp")) {
-    if (argc != 3) {
-      gtty->Printf("invalid argument\n");
-      return;
-    }
-
-    DhcpCtrl::GetCtrl().AssignAddr(dev);
-  } else {
-    gtty->Printf("invalid argument\n");
-    return;
-  }
-}
-
-static void ifconfig(int argc, const char *argv[]) {
-  uptr<Array<const char *>> list = netdev_ctrl->GetNamesOfAllDevices();
-  gtty->Printf("\n");
-  for (size_t l = 0; l < list->GetLen(); l++) {
-    gtty->Printf("%s", (*list)[l]);
-    NetDev *dev = netdev_ctrl->GetDeviceInfo((*list)[l])->device;
-    dev->UpdateLinkStatus();
-    gtty->Printf("  link: %s\n", dev->IsLinkUp() ? "up" : "down");
-  }
-}
-
-void setup_arp_reply(NetDev *dev);
-void send_arp_packet(NetDev *dev, uint8_t *ipaddr);
-
-static void bench(int argc, const char *argv[]) {
-  if (argc == 1) {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-  if (strcmp(argv[1], "snd") == 0) {
-    if (argc == 2) {
-      gtty->Printf("specify ethernet interface.\n");
-      return;
-    } else if (argc == 3) {
-      gtty->Printf("specify ip v4 addr.\n");
-      return;
-    } else if (argc != 4) {
-      gtty->Printf("invalid arguments\n");
-      return;
-    }
-    NetDevCtrl::NetDevInfo *info = netdev_ctrl->GetDeviceInfo(argv[2]);
-    if (info == nullptr) {
-      gtty->Printf("no ethernet interface(%s).\n", argv[2]);
-      return;
-    }
-    NetDev *dev = info->device;
-
-    uint8_t addr[4] = {0, 0, 0, 0};
-    if (!parse_ipaddr(argv[3], addr)) {
-      gtty->Printf("invalid ip v4 addr.\n");
-      return;
-    }
-
-    send_arp_packet(dev, addr);
-  } else if (strcmp(argv[1], "set_reply") == 0) {
-    if (argc == 2) {
-      gtty->Printf("specify ethernet interface.\n");
-      return;
-    }
-    NetDevCtrl::NetDevInfo *info = netdev_ctrl->GetDeviceInfo(argv[2]);
-    if (info == nullptr) {
-      gtty->Printf("no ethernet interface(%s).\n", argv[2]);
-      return;
-    }
-    setup_arp_reply(info->device);
-  } /* else if (strcmp(argv[1], "qemu") == 0) {
-    uint8_t ip1_[] = {10, 0, 2, 9};
-    uint8_t ip2_[] = {10, 0, 2, 15};
-    memcpy(ip1, ip1_, 4);
-    memcpy(ip2, ip2_, 4);
-    sdevice = (argc == 2) ? "eth0" : argv[2];
-    rdevice = (argc == 2) ? "eth0" : argv[2];
-    if (!netdev_ctrl->Exists(rdevice)) {
-      gtty->Printf("no ethernet interface(%s).\n", rdevice);
-      return;
-    }
-    } */ else {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-}
-
-static void setflag(int argc, const char *argv[]) {
-  if (argc == 1) {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-  if (strcmp(argv[1], "spinlock_timeout") == 0) {
-    if (argc == 2) {
-      gtty->Printf("invalid argument.\n");
-      return;
-    }
-    if (strcmp(argv[2], "true") == 0) {
-      SpinLock::_spinlock_timeout = true;
-    } else if (strcmp(argv[2], "false") == 0) {
-      SpinLock::_spinlock_timeout = false;
-    } else {
-      gtty->Printf("invalid argument.\n");
-      return;
-    }
-  } else {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-}
-
-static void udp_setup(int argc, const char *argv[]) {
-  UdpCtrl::GetCtrl().SetupServer();
-}
-
-static void arp_scan(int argc, const char *argv[]) {
-  auto devices = netdev_ctrl->GetNamesOfAllDevices();
-  for (size_t i = 0; i < devices->GetLen(); i++) {
-    auto dev = netdev_ctrl->GetDeviceInfo((*devices)[i])->device;
-    if (dev->GetStatus() != NetDev::LinkStatus::kUp) {
-      gtty->Printf("skip %s (Link Down)\n", (*devices)[i]);
-      continue;
-    }
-    uint32_t my_addr_int_;
-    if (!dev->GetIpv4Address(my_addr_int_) || my_addr_int_ == 0xFFFFFFFF) {
-      gtty->Printf("skip %s (no IP)\n", (*devices)[i]);
-      continue;
-    }
-    uint8_t ipaddr[4];
-    *(reinterpret_cast<uint32_t *>(ipaddr)) = my_addr_int_;
-    gtty->Printf("ARP scan with %s(%d.%d.%d.%d)\n", (*devices)[i], ipaddr[0], ipaddr[1], ipaddr[2], ipaddr[3]);
-    dev->SetReceiveCallback(
-        network_cpu,
-        make_uptr(new Function<NetDev *>(
-            [](NetDev *eth) {
-              NetDev::Packet *rpacket;
-              if (!eth->ReceivePacket(rpacket)) {
-                return;
-              }
-              uint32_t my_addr_int;
-              assert(eth->GetIpv4Address(my_addr_int));
-              uint8_t my_addr[4];
-              *(reinterpret_cast<uint32_t *>(my_addr)) = my_addr_int;
-              // received packet
-              if (rpacket->GetBuffer()[12] == 0x08 &&
-                  rpacket->GetBuffer()[13] == 0x06 &&
-                  rpacket->GetBuffer()[21] == 0x02) {
-                // ARP Reply
-                uint32_t target_addr_int = (rpacket->GetBuffer()[31] << 24) |
-                                           (rpacket->GetBuffer()[30] << 16) |
-                                           (rpacket->GetBuffer()[29] << 8) |
-                                           rpacket->GetBuffer()[28];
-                arp_table->Set(target_addr_int, rpacket->GetBuffer() + 22, eth);
-              }
-              if (rpacket->GetBuffer()[12] == 0x08 &&
-                  rpacket->GetBuffer()[13] == 0x06 &&
-                  rpacket->GetBuffer()[21] == 0x01 &&
-                  (memcmp(rpacket->GetBuffer() + 38, my_addr, 4) == 0)) {
-                // ARP Request
-                uint8_t data[] = {
-                    0x00, 0x00, 0x00, 0x00, 0x00, 0x00,  // Target MAC Address
-                    0x00, 0x00, 0x00, 0x00, 0x00, 0x00,  // Source MAC Address
-                    0x08, 0x06,                          // Type: ARP
-                    // ARP Packet
-                    0x00, 0x01,  // HardwareType: Ethernet
-                    0x08, 0x00,  // ProtocolType: IPv4
-                    0x06,        // HardwareLength
-                    0x04,        // ProtocolLength
-                    0x00, 0x02,  // Operation: ARP Reply
-                    0x00, 0x00, 0x00, 0x00, 0x00,
-                    0x00,                    // Source Hardware Address
-                    0x00, 0x00, 0x00, 0x00,  // Source Protocol Address
-                    0x00, 0x00, 0x00, 0x00, 0x00,
-                    0x00,                    // Target Hardware Address
-                    0x00, 0x00, 0x00, 0x00,  // Target Protocol Address
-                };
-                memcpy(data, rpacket->GetBuffer() + 6, 6);
-                static_cast<DevEthernet *>(eth)->GetEthAddr(data + 6);
-                memcpy(data + 22, data + 6, 6);
-                memcpy(data + 28, my_addr, 4);
-                memcpy(data + 32, rpacket->GetBuffer() + 22, 6);
-                memcpy(data + 38, rpacket->GetBuffer() + 28, 4);
-
-                uint32_t len = sizeof(data) / sizeof(uint8_t);
-                NetDev::Packet *tpacket;
-                kassert(eth->GetTxPacket(tpacket));
-                memcpy(tpacket->GetBuffer(), data, len);
-                tpacket->len = len;
-                eth->TransmitPacket(tpacket);
-              }
-              eth->ReuseRxBuffer(rpacket);
-            },
-            dev)));
-    for (int j = 1; j < 255; j++) {
-      uint32_t my_addr_int;
-      assert(dev->GetIpv4Address(my_addr_int));
-      uptr<Thread> thread = ThreadCtrl::GetCtrl(cpu_ctrl->RetainCpuIdForPurpose(CpuPurpose::kLowPriority)).AllocNewThread(Thread::StackState::kShared);
-      do {
-        auto t_op = thread->CreateOperator();
-        struct Container {
-          NetDev *eth;
-          uint8_t target_addr[4];
-        };
-        auto container_ = make_uptr(new Container);
-        container_->eth = dev;
-        *(reinterpret_cast<uint32_t *>(container_->target_addr)) = my_addr_int;
-        container_->target_addr[3] = j;
-        t_op.SetFunc(make_uptr(new Function<uptr<Container>>([](uptr<Container> container) {
-                uint8_t data[] = {
-                  0xff, 0xff, 0xff, 0xff, 0xff, 0xff,  // Target MAC Address
-                  0x00, 0x00, 0x00, 0x00, 0x00, 0x00,  // Source MAC Address
-                  0x08, 0x06,                          // Type: ARP
-                  // ARP Packet
-                  0x00, 0x01,                          // HardwareType: Ethernet
-                  0x08, 0x00,                          // ProtocolType: IPv4
-                  0x06,                                // HardwareLength
-                  0x04,                                // ProtocolLength
-                  0x00, 0x01,                          // Operation: ARP Request
-                  0x00, 0x00, 0x00, 0x00, 0x00, 0x00,  // Source Hardware Address
-                  0x00, 0x00, 0x00, 0x00,              // Source Protocol Address
-                  0x00, 0x00, 0x00, 0x00, 0x00, 0x00,  // Target Hardware Address
-                  // Target Protocol Address
-                  0x00, 0x00, 0x00, 0x00};
-                static_cast<DevEthernet *>(container->eth)->GetEthAddr(data + 6);
-                memcpy(data + 22, data + 6, 6);
-                uint32_t my_addr;
-                assert(container->eth->GetIpv4Address(my_addr));
-                *(reinterpret_cast<uint32_t *>(data)) = my_addr;
-                memcpy(data + 38, container->target_addr, 4);
-                uint32_t len = sizeof(data) / sizeof(uint8_t);
-                NetDev::Packet *tpacket;
-                kassert(container->eth->GetTxPacket(tpacket));
-                memcpy(tpacket->GetBuffer(), data, len);
-                tpacket->len = len;
-                container->eth->TransmitPacket(tpacket);
-              },
-              container_)));
-        t_op.Schedule();
-      } while(0);
-      thread->Join();
-    }
-  }
-  uptr<Thread> thread = ThreadCtrl::GetCtrl(cpu_ctrl->RetainCpuIdForPurpose(CpuPurpose::kLowPriority)).AllocNewThread(Thread::StackState::kShared);
-  do {
-    auto t_op = thread->CreateOperator();
-    t_op.SetFunc(make_uptr(new Function<void *>([](void *) {
-            auto devices_ = netdev_ctrl->GetNamesOfAllDevices();
-            for (size_t i = 0; i < devices_->GetLen(); i++) {
-              auto dev = netdev_ctrl->GetDeviceInfo((*devices_)[i])->device;
-              dev->SetReceiveCallback(network_cpu,
-                                      make_uptr(new GenericFunction<>()));
-            }
-          }, nullptr)));
-    t_op.Schedule(3 * 1000 * 1000);
-  } while(0);
-  thread->Join();
-}
-
 ArpTable *arp_table = nullptr;
 
-void udpsend(int argc, const char *argv[]) {
-  if (argc != 4) {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-
-  uint8_t target_addr[4] = {0, 0, 0, 0};
-  if (!parse_ipaddr(argv[1], target_addr)) {
-    gtty->Printf("invalid ip v4 addr.\n");
-    return;
-  }
-  UdpCtrl::GetCtrl().SendStr(&target_addr, atoi(argv[2]), argv[3]);
-}
-
-static void show(int argc, const char *argv[]) {
-  if (argc == 1) {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-  if (strcmp(argv[1], "module") == 0) {
-    if (argc != 2) {
-      gtty->Printf("invalid argument.\n");
-      return;
-    }
-    multiboot_ctrl->ShowModuleInfo();
-  } else if (strcmp(argv[1], "info") == 0) {
-    gtty->Printf("[kernel] info: Hardware Information\n");
-    gtty->Printf("available cpu thread num: %d\n", cpu_ctrl->GetHowManyCpus());
-    gtty->Printf("\n");
-    gtty->Printf("[kernel] info: Build Information\n");
-    multiboot_ctrl->ShowBuildTimeStamp();
-  } else {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-}
-
-static void wait_until_linkup(int argc, const char *argv[]) {
-  if (argc != 2) {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-  NetDevCtrl::NetDevInfo *info = netdev_ctrl->GetDeviceInfo(argv[1]);
-  if (info == nullptr) {
-    gtty->Printf("no such device.\n");
-    return;
-  }
-  NetDev *dev_ = info->device;
-  uptr<Thread> thread = ThreadCtrl::GetCtrl(cpu_ctrl->RetainCpuIdForPurpose(CpuPurpose::kLowPriority)).AllocNewThread(Thread::StackState::kShared);
-  do {
-    Thread::Operator t_op = thread->CreateOperator();
-    t_op.SetFunc(make_uptr(new Function<NetDev *>([](NetDev *dev) {
-            if (!dev->IsLinkUp()) {
-              ThreadCtrl::GetCurrentThreadOperator().Schedule(1000 * 1000);
-            }
-          }, dev_)));
-    t_op.Schedule();
-  } while(0);
-  thread->Join();
-}
-
-static void load_script(uptr<Array<uint8_t>> data) {
-  size_t old_index = 0;
-  size_t index = 0;
-  while (index <= data->GetLen()) {
-    if (index == data->GetLen() ||
-        (*data)[index] == '\r' ||
-        (*data)[index] == '\n' ||
-         (*data)[index] == '\0') {
-      if (old_index != index) {
-        char buffer[index - old_index + 1];
-        memcpy(buffer,
-               reinterpret_cast<char *>(data->GetRawPtr()) + old_index,
-               index - old_index);
-        buffer[index - old_index] = '\0';
-        auto ec = make_uptr(new Shell::ExecContainer(shell));
-        ec = shell->Tokenize(ec, buffer);
-        int timeout = 0;
-        if (strlen(buffer) != 0) {
-          gtty->Printf("> %s\n", buffer);
-          if (strcmp(ec->argv[0], "wait") == 0) {
-            if (ec->argc == 2) {
-              int t = 0;
-              for (size_t l = 0; l < strlen(ec->argv[1]); l++) {
-                if ('0' > ec->argv[1][l] || ec->argv[1][l] > '9') {
-                  gtty->Printf("invalid argument.\n");
-                  t = 0;
-                  break;
-                }
-                t = t * 10 + ec->argv[1][l] - '0';
-              }
-              timeout = t * 1000 * 1000;
-            } else {
-              gtty->Printf("invalid argument.\n");
-            }
-          } else if (strcmp(ec->argv[0], "wait_until_linkup") == 0) {
-            wait_until_linkup(ec->argc, ec->argv);
-          } else {
-            shell->Execute(ec);
-          }
-        }
-        if (timeout != 0) {
-          ThreadCtrl::GetCurrentThreadOperator().Schedule(timeout);
-          ThreadCtrl::WaitCurrentThread();
-        }
-      }
-      if (index == data->GetLen() - 1 || index == data->GetLen() || (*data)[index] == '\0') {
-        break;
-      }
-      old_index = index + 1;
-    }
-    index++;
-  }
-}
-
-static void load(int argc, const char *argv[]) {
-  if (argc != 2) {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-  if (strcmp(argv[1], "script.sh") == 0) {
-    load_script(multiboot_ctrl->LoadFile(argv[1]));
-  } else if (strcmp(argv[1], "test.elf") == 0) {
-    auto buf_ = multiboot_ctrl->LoadFile(argv[1]);
-    uptr<Thread> thread = ThreadCtrl::GetCtrl(cpu_ctrl->RetainCpuIdForPurpose(CpuPurpose::kLowPriority)).AllocNewThread(Thread::StackState::kIndependent);
-    do {
-      auto t_op = thread->CreateOperator();
-      t_op.SetFunc(make_uptr(new Function<uptr<Array<uint8_t>>>([](uptr<Array<uint8_t>> buf) {
-              Loader loader;
-              ElfObject obj(loader, buf->GetRawPtr());
-              if (obj.Init() != BinObjectInterface::ErrorState::kOk) {
-                gtty->Printf("error while loading app\n");
-              } else {
-                obj.Execute();
-              }
-            }, buf_)));
-      t_op.Schedule();
-    } while(0);
-    thread->Join();
-  } else if (strcmp(argv[1], "rump.bin") == 0) {
-    auto buf_ = multiboot_ctrl->LoadFile(argv[1]);
-    uptr<Thread> thread = ThreadCtrl::GetCtrl(cpu_ctrl->RetainCpuIdForPurpose(CpuPurpose::kLowPriority)).AllocNewThread(Thread::StackState::kIndependent);
-    do {
-      auto t_op = thread->CreateOperator();
-      t_op.SetFunc(make_uptr(new Function<uptr<Array<uint8_t>>>([](uptr<Array<uint8_t>> buf) {
-              Ring0Loader loader;
-              RaphineRing0AppObject obj(loader, buf->GetRawPtr());
-              if (obj.Init() != BinObjectInterface::ErrorState::kOk) {
-                gtty->Printf("error while loading app\n");
-              } else {
-                obj.Execute();
-              }
-            }, buf_)));
-      t_op.Schedule();
-    } while(0);
-    thread->Join();
-  } else {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-}
-
-static void remote_load(int argc, const char *argv[]) {
-  if (argc != 2) {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-  if (strcmp(argv[1], "test.elf") == 0) {
-    SystemCallCtrl::GetCtrl().SetMode(SystemCallCtrl::Mode::kRemote);
-    auto buf_ = multiboot_ctrl->LoadFile(argv[1]);
-    uptr<Thread> thread = ThreadCtrl::GetCtrl(cpu_ctrl->RetainCpuIdForPurpose(CpuPurpose::kLowPriority)).AllocNewThread(Thread::StackState::kShared);
-    do {
-      auto t_op = thread->CreateOperator();
-      t_op.SetFunc(make_uptr(new Function<uptr<Array<uint8_t>>>([](uptr<Array<uint8_t>> buf) {
-              Loader loader;
-              ElfObject obj(loader, buf->GetRawPtr());
-              if (obj.Init() != BinObjectInterface::ErrorState::kOk) {
-                gtty->Printf("error while loading app\n");
-              } else {
-                obj.Execute();
-              }
-            }, buf_)));
-      t_op.Schedule();
-    } while(0);
-    thread->Join();
-  } else {
-    gtty->Printf("invalid argument.\n");
-    return;
-  }
-}
-
-void beep(int argc, const char *argv[]) {
-  uptr<Thread> thread = ThreadCtrl::GetCtrl(cpu_ctrl->RetainCpuIdForPurpose(CpuPurpose::kLowPriority)).AllocNewThread(Thread::StackState::kShared);
-  do {
-    auto t_op = thread->CreateOperator();
-    t_op.SetFunc(make_uptr(new Function<void *>([](void *) {
-            static int i = 0;
-            if (i < 6) {
-              uint16_t sound[6] = {905, 761, 452, 570, 508, 380};
-              uint8_t a = 0xb6;
-              outb(0x43, a);
-              uint8_t l = sound[i] & 0x00FF;
-              outb(0x42, l);
-              uint8_t h = (sound[i] >> 8) & 0x00FF;
-              outb(0x42, h);
-              uint8_t on = inb(0x61);
-              outb(0x61, (on | 0x03) & 0x0f);
-              i++;
-              ThreadCtrl::GetCurrentThreadOperator().Schedule(110000);
-            } else {
-              uint8_t off = inb(0x61);
-              outb(0x61, off & 0xd);
-            }
-          }, nullptr)));
-    t_op.Schedule();
-  } while(0);
-  thread->Join();
-}
-
-static void membench(int argc, const char *argv[]) {
-  register_membench2_callout();
-}
-
-bool cat_sub(VirtualFileSystem *vfs, const char *path) {
-  InodeContainer inode;
-  if (vfs->LookupInodeFromPath(inode, path, false) != IoReturnState::kOk)
-    return false;
-  VirtualFileSystem::Stat st;
-  if (inode.GetStatOfInode(st) != IoReturnState::kOk) return false;
-  size_t s = st.size;
-  uint8_t buf[s];
-  if (inode.ReadData(buf, 0, s) != IoReturnState::kOk) return false;
-  for (size_t i = 0; i < s; i++) {
-    gtty->Printf("%c", buf[i]);
-  }
-  gtty->Printf("\n");
-  return true;
-}
-
-void cat(int argc, const char *argv[]) {
-  if (argc < 2) {
-    gtty->Printf("usage: cat <filename>\n");
-    return;
-  }
-  Storage *storage;
-  if (StorageCtrl::GetCtrl().GetDevice("ram0", storage) != IoReturnState::kOk) {
-    kernel_panic("storage", "not found");
-  }
-  V6FileSystem *v6fs = new V6FileSystem(*storage);
-  VirtualFileSystem *vfs = new VirtualFileSystem(v6fs);
-  vfs->Init();
-  if (!cat_sub(vfs, argv[1])) {
-    gtty->Printf("cat: %s: No such file or directory\n", argv[1]);
-  }
-}
+void RegisterDefaultShellCommands();
+void load_script(uptr<Array<uint8_t>> data);
 
 void freebsd_main();
-
 extern "C" int main() {
   _system_memory_space.GetKernelVirtmemCtrl()->Init();
 
@@ -728,7 +127,7 @@ extern "C" int main() {
   // arp_table = new (&_arp_table) ArpTable();
 
   physmem_ctrl->Init();
-  
+
   multiboot_ctrl->Setup();
 
   system_memory_space->GetKernelVirtmemCtrl()->InitKernelMemorySpace();
@@ -791,7 +190,7 @@ extern "C" int main() {
   freebsd_main();
 
   AttachDevices<PciCtrl, LegacyKeyboard, Ramdisk, Device>();
-  
+
   SystemCallCtrl::Init();
 
   gtty->Printf("\n\n[kernel] info: initialization completed\n");
@@ -801,28 +200,15 @@ extern "C" int main() {
   DhcpCtrl::Init();
 
   shell->Setup();
-  shell->Register("halt", halt);
-  shell->Register("reset", reset);
-  shell->Register("bench", bench);
-  shell->Register("beep", beep);
-  shell->Register("lspci", lspci);
-  shell->Register("ifconfig", ifconfig);
-  shell->Register("setip", setip);
-  shell->Register("show", show);
-  shell->Register("load", load);
-  shell->Register("remote_load", remote_load);
-  shell->Register("setflag", setflag);
-  shell->Register("udpsend", udpsend);
-  shell->Register("arp_scan", arp_scan);
-  shell->Register("udp_setup", udp_setup);
-  shell->Register("membench", membench);
-  shell->Register("cat", cat);
+  RegisterDefaultShellCommands();
 
-  auto thread = ThreadCtrl::GetCtrl(cpu_ctrl->RetainCpuIdForPurpose(CpuPurpose::kLowPriority)).AllocNewThread(Thread::StackState::kIndependent);
+  auto thread = ThreadCtrl::GetCtrl(
+                    cpu_ctrl->RetainCpuIdForPurpose(CpuPurpose::kLowPriority))
+                    .AllocNewThread(Thread::StackState::kIndependent);
   auto t_op = thread->CreateOperator();
-  t_op.SetFunc(make_uptr(new Function<void *>([](void *) {
-          load_script(multiboot_ctrl->LoadFile("init.sh"));
-        }, nullptr)));
+  t_op.SetFunc(make_uptr(new Function<void *>(
+      [](void *) { load_script(multiboot_ctrl->LoadFile("init.sh")); },
+      nullptr)));
   t_op.Schedule();
 
   ThreadCtrl::GetCurrentCtrl().Run();
